requirements.txt
pandas
selenium
python-dotenv